pyproject.toml
[tool.poetry]
name = "ppo"
version = "0.1.0"
description = ""
authors = ["Logan Walls <2934282+LoganWalls@users.noreply.github.com>", "Ethan Brooks <ethanabrooks@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
jaxlib = "0.3.0"
jax = "0.3.6"
gym-minigrid = "^1.1.0"
dollar-lambda = "^1.1.4"
run-logger = "^0.1.8"
tensorflow-gpu = "^2.9.1"
flax = "^0.5.3"
GitPython = "^3.1.27"
requests-toolbelt = "^0.9.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
ipdb = "^0.13.9"
ipython = "^8.4.0"
pre-commit = "^2.20.0"

[tool.poetry.scripts]
main = 'ppo:ppo_main.main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
